\docType{class}
\name{adjust_lmList-class}
\alias{adjust_lmList-class}
\alias{coef,adjust_lmList-method}
\alias{confint,adjust_lmList-method}
\alias{formula,adjust_lmList-method}
\alias{plot,adjust_lmList.confint,ANY-method}
\alias{show,adjust_lmList-method}
\title{Class \code{"adjust_lmList"}}
\description{
  Adjusted List of \code{lm} Objects with a Common Model
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{new("adjust_lmList", ...)} or by
  \code{adjust_lmList}.
}
\keyword{classes}