% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm_augment.R
\name{hlm_augment.default}
\alias{hlm_augment.default}
\title{Calculating residuals and influence diagnostics for HLMs}
\usage{
\method{hlm_augment}{default}(object, ...)
}
\description{
Calculating residuals and influence diagnostics for HLMs
}
