% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm_influence.R
\name{hlm_influence.default}
\alias{hlm_influence.default}
\alias{hlm_influence.lmerMod}
\alias{hlm_influence}
\title{Calculating influence diagnostics for HLMs}
\usage{
\method{hlm_influence}{default}(model, ...)

\method{hlm_influence}{lmerMod}(model, level = 1, approx = TRUE, leverage = "overall", ...)
}
\arguments{
\item{model}{an object of class \code{lmerMod}}

\item{level}{a variable used to define the group for which cases are deleted and influence 
diagnostics are calculated. If \code{level} equals 1 (default), then influence diagnostics are 
calculated for individual observations. Otherwise, \code{level} should be the name of a grouping
factor as defined in \code{flist} of the \code{lmerMod} object.}

\item{approx}{logical parameter used to determine how the influence diagnostics are calculated.
If \code{FALSE} (default), influence diagnostics are calculated using a one step approximation. 
If \code{TRUE}, influence diagnostics are caclulated by iteratively deleting groups and refitting
the model using \code{lmer}. This method is more accurate, but slower than the one step approximation. 
If \code{approx = FALSE}, the returned tibble also contains columns for relative variance change (RVC).}

\item{leverage}{a character vector to determine which types of leverage should be included in the 
returned tibble. There are four options: 'overall' (default), 'fixef', 'ranef', or 'ranef.uc'. 
One or more types may be specified. For additional information about the types of leverage, see 
\code{?leverage}.}
}
\description{
This function is used to compute influence diagnostics for a hierarchical linear model. 
It takes a model fit as a \code{lmerMod} object and returns a tibble with Cook's 
distance, MDFFITS, covtrace, covratio, and leverage.
}
\details{
The \code{hlm_influence} function provides a wrapper that appends influence diagnostics 
to the original data. The approximated influence diagnostics returned by this 
function are equivalent to those returned by \code{cooks.distance}, \code{mdffits}, \code{covtrace}, 
\code{covratio}, and \code{leverage}. The exact influence diagnostics obtained through a full 
refit of the data are also avaliable through \code{case_delete} and the accompaning functions 
\code{cooks.distance}, \code{mdffits}, \code{covtrace}, and \code{covratio} that can be called 
directly on the \code{case_delete} object.
}
