% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLMaugment.R
\name{HLMaugment.default}
\alias{HLMaugment.default}
\alias{HLMaugment.lmerMod}
\alias{HLMaugment}
\title{Calculating residuals from HLMs}
\usage{
\method{HLMaugment}{default}(object, ...)

\method{HLMaugment}{lmerMod}(object, level = 1, standardize = FALSE, sim = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{lmerMod}.}

\item{...}{do not use}

\item{level}{which residuals should be extracted: 1 for within-group
(case-level) residuals, the name of a grouping factor (as defined in
\code{flist} of the \code{lmerMod} object) for between-group residuals}

\item{standardize}{if \code{standardize = TRUE} the standardized residuals
will be returned; if \code{standardize = "semi"} then the semi-standardized
level-1 residuals will be returned}

\item{sim}{optional argument giving the data frame used for LS residuals.
This is used mainly for dealing with simulations.}
}
\description{
\code{HLMaugment} takes a hierarchical linear model fit as a
\code{lmerMod} object and adds information about each observation's 
residuals and predicted values.
}
\details{

}
