% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLMaugment.R
\name{HLMaugment}
\alias{HLMaugment}
\alias{HLMresid}
\title{Calculating residuals from HLMs}
\usage{
\method{HLMresid}{mer}(object, level = 1, standardize = FALSE, sim = NULL)
}
\arguments{
\item{object}{an object of class \code{lmerMod}.}

\item{standardize}{if \code{standardize = TRUE} the standardized
residuals will be returned; if \code{standardize = "semi"} then
the semi-standardized level-1 residuals will be returned.}

\item{sim}{optional argument giving the data frame used for LS residuals. This
is used mainly for dealing with simulations.}

\item{...}{do not use}
}
\description{
\code{HLMaugment} is a function that extracts residuals
from a hierarchical linear model fit
using \code{lmer}. That is, it is a unified framework that
extracts/calculates residuals from \code{mer} or \code{lmerMod} objects.
}
\details{
This function extracts residuals from the model, 
using least squares (LS) and Empirical 
Bayes (EB) methods. This unified framework
enables the analyst to more easily conduct
an upward residual analysis during model
exploration/checking.

The \code{HLMaugment} function provides a wrapper that will extract
residuals from a fitted \code{mer} or \code{lmerMod} object.
}
