% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLMdiag-defunct.R
\name{HLMdiag-defunct}
\alias{HLMdiag-defunct}
\alias{HLMresid}
\alias{HLMresid.default}
\alias{HLMresid.lmerMod}
\alias{HLMresid.mer}
\alias{diagnostics}
\alias{group_qqnorm}
\alias{ggplot_qqnorm}
\title{Defunct functions in package HLMdiag}
\usage{
HLMresid(...)

HLMresid.default(...)

HLMresid.lmerMod(...)

HLMresid.mer(...)

diagnostics(...)

group_qqnorm(...)

ggplot_qqnorm(...)
}
\arguments{
\item{...}{arguments passed to defunct functions}
}
\description{
These functions are defunct and no longer available.
}
\details{
\code{HLMresid} is replaced by \code{\link{hlm_resid}}

\code{diagnostics} is replaced by \code{\link{hlm_influence}}

\code{group_qqnorm} and \code{group_qqnorm} are replaced by functions in \pkg{qqplotr}. 
   See \code{\link[qqplotr]{stat_qq_point}}, \code{\link[qqplotr]{stat_qq_line}}, and 
   \code{\link[qqplotr]{stat_qq_band}}.
}
