% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLMdiag-deprecated.R
\name{HLMdiag-deprecated}
\alias{HLMdiag-deprecated}
\title{Deprecated functions in HLMdiag}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
 \item \code{\link{HLMresid}}: This function is deprecated, and will
 be removed in the next version of this package.
 \item \code{\link{diagnostics}}: This function is deprecated, and will
 be removed in the next version of this package.
}
}
