\docType{package}
\name{HLMdiag}
\alias{HLMdiag}
\alias{HLMdiag-package}
\alias{package-HLMdiag}
\title{Diagnostic tools for hierarchical (multilevel) linear models}
\description{
  A suite of diagnostic tools for hierarchical (multilevel)
  linear models. The package offers not only leverage and
  traditional deletion diagnostics (Cook's distance,
  covratio, covtrace, and MDFFITS) but also provides
  convenience functions and graphics for residual analysis.
  The package assumes that models were fit using the lme4
  package.
}
\keyword{package}

