\name{leverage.mer}
\alias{leverage}
\alias{leverage.mer}
\title{Leverage for HLMs}
\usage{
  \method{leverage}{mer} (object, level, ...)
}
\arguments{
  \item{object}{fitted object of class \code{mer}}

  \item{level}{the level at which the leverage should be
  calculated: either 1 for observation level leverage or
  the name of the grouping factor (as defined in
  \code{flist} of the \code{mer} object) for group level
  leverage. \code{leverage} assumes that the grouping
  factors are unique; thus, if IDs are repeated within each
  unit, unique IDs must be generated by the user prior to
  use of \code{leverage}.}

  \item{...}{do not use}
}
\description{
  This function calculates the leverage of a hierarchical
  linear model fit by \code{lmer}.
}
\examples{
data(sleepstudy, package = 'lme4')
fm <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)

# Observation level leverage
lev1 <- leverage.mer(fm, level = 1)
head(lev1)
}
\author{
  Adam Loy \email{aloy@iastate.edu}
}
\references{
  Demidenko, E., & Stukel, T. A. (2005) Influence analysis
  for linear mixed-effects models.  \emph{Statistics in
  Medicine}, \bold{24}(6), 893--909.
}
\seealso{
  \code{\link{cooks.distance.mer}},
  \code{\link{mdffits.mer}}, \code{\link{covratio.mer}},
  \code{\link{covtrace.mer}}, \code{\link{rvc.mer}}
}
\keyword{models}
\keyword{regression}

