\name{pplineInfo}
\alias{pplineInfo}
\title{Adding a line to a normal probability plot}
\usage{
pplineInfo(x)
}
\arguments{
  \item{x}{a numeric vector}
}
\description{
  This function will calculate the equation of the line
  that can be used to help read the normal probability plot
  of interest.
}
\author{
  Adam Loy \email{aloy@istate.edu}
}
