\name{qqlineInfo}
\alias{qqlineInfo}
\title{Adding a line to a normal quantile-quantile plot}
\usage{
qqlineInfo(x)
}
\arguments{
  \item{x}{a numeric vector}
}
\description{
  This function will calculate the equation of the line
  that can be used to help read the normal quantile plot of
  interest.
}
\author{
  Adam Loy \email{aloy@istate.edu}
}
