% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_functions.R
\name{resid_ranef}
\alias{resid_ranef}
\title{Random effects residuals}
\usage{
resid_ranef(object, level, which, standardize)
}
\arguments{
\item{object}{an object of class \code{lmerMod}.}

\item{level}{DESCRIPTION}

\item{which}{DESCRIPTION}

\item{standardize}{DESCRIPTION}
}
\value{
A vector of conditional residuals.
}
\description{
Calculates Random effects  residuals of \code{lmerMod} model objects.
}
