\name{sampleQuantiles}
\alias{sampleQuantiles}
\title{Calculating quantiles...}
\usage{sampleQuantiles(x)}
\description{Calculating quantiles}
\details{This function will summarize the quantiles that are found within the
data vector provided.}
\author{Adam Loy \email{aloy@iastate.edu}}
\arguments{\item{x}{a numeric vector}}
