% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm_resid.R
\name{hlm_resid.default}
\alias{hlm_resid.default}
\alias{hlm_resid.lmerMod}
\alias{hlm_resid}
\title{Calculating residuals from HLMs}
\usage{
\method{hlm_resid}{default}(object, ...)

\method{hlm_resid}{lmerMod}(object, level = 1, standardize = FALSE, sim = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{lmerMod}.}

\item{...}{do not use}

\item{level}{which residuals should be extracted: 1 for within-group
(case-level) residuals, the name of a grouping factor (as defined in
\code{flist} of the \code{lmerMod} object) for between-group residuals}

\item{standardize}{if \code{standardize = TRUE} the standardized residuals
will be returned; if \code{standardize = "semi"} then the semi-standardized
level-1 residuals will be returned}

\item{sim}{optional argument giving the data frame used for LS residuals.
This is used mainly for dealing with simulations.}
}
\description{
\code{hlm_resid} takes a hierarchical linear model fit as a
\code{lmerMod} object and adds information about each observation's 
residuals and predicted values.
}
\details{
This function extract residuals and predicted values from the model, using
least squares (LS) and Empirical Bayes (EB) methods, and appends them to the
model data. This unified framework enables the analyst to more easily conduct
an upward residual analysis during model exploration/checking.

The \code{hlm_resid} function provides a wrapper that will extract
residuals and predicted values from a fitted \code{lmerMod} object. 
The function provides access to 
residual quantities already made available by the functions \code{resid},
\code{predict}, and \code{ranef}, but adds additional functionality. Below is
a list of types of residuals and predicted values that are extracted and
appended to the model data.
\describe{
\item{raw level-1 LS residuals}{These are equivalent to the residuals extracted
by \code{resid} if \code{level = 1}, \code{type = "EB"}, and 
\code{standardize = FALSE} is specified. }
\item{level-1 LS fitted values}{The predicted values }
}
Note that \code{standardize = "semi"} is only implemented for level-1 LS residuals.
}
