\name{identify_resid}
\alias{identify_resid}
\title{Identifying unusual points...}
\usage{identify_resid(formula, identify)}
\description{Identifying unusual points}
\details{This function will identify the unusual points found when
looking at the plot comparing the shrinkage estimates and
the random coefficients found through OLS.}
\author{Adam Loy \email{aloy@iastate.edu}}
\arguments{\item{formula}{a formula that can be used with the \code{lm} function.}
\item{identify}{the percentage (decimal form) of points to identify as unusual}}
