\name{leverage.mer}
\alias{leverage}
\alias{leverage.mer}
\title{Leverage for mixed/hierarchical linear models}
\usage{
  \method{leverage}{mer} (object, level, ...)
}
\arguments{
  \item{object}{fitted object of class \code{mer}}

  \item{level}{the level at which the leverage should be
  calculated; either 1 for observation level leverage or
  the name of the grouping factor (as defined in
  \code{flist} of the \code{mer} object) for group level
  leverage. \code{leverage} assumes that the grouping
  factors are unique; thus, if IDs are repeated within each
  unit, unique IDs must be generated by the user prior to
  use of \code{leverage}.}

  \item{...}{do not use}
}
\description{
  This function calculates the leverage of a
  mixed/hierarchical model fit by \code{lmer}.
}
\author{
  Adam Loy \email{aloy@iastate.edu}
}
\references{
  Nobre, J. S., & Singer, J. M. (2011).  Leverage analysis
  for linear mixed models.  \emph{Journal of Applied
  Statistics}, 38(5), 1063--1072.

  Demidenko, E., & Stukel, T. A. (2005).  Influence
  analysis for linear mixed-effects models.
  \emph{Statistics in Medicine}, 24(6), 893--909.
}
\seealso{
  \code{\link{cooks.distance.mer}},
  \code{\link{mdffits.mer}}, \code{\link{covratio.mer}},
  \code{\link{covtrace.mer}}, \code{\link{rvc.mer}}
}

