\name{pplineInfo}
\alias{pplineInfo}
\title{Adding a line to a normal probability plot...}
\usage{pplineInfo(x)}
\description{Adding a line to a normal probability plot}
\details{This function will calculate the equation of the line fit through the
the first and third quartiles that can be used to help read the normal probability plot of interest.}
\author{Adam Loy \email{aloy@iastate.edu}}
\arguments{\item{x}{a numeric vector}}