\name{qqlineInfo}
\alias{qqlineInfo}
\title{Adding a line to a normal quantile-quantile plot...}
\usage{qqlineInfo(x)}
\description{Adding a line to a normal quantile-quantile plot}
\details{This function will calculate the equation of the line  fit through the the first and third quartiles that can be
used to help read the normal quantile plot of interest.}
\author{Adam Loy \email{aloy@iastate.edu}}
\arguments{\item{x}{a numeric vector}}